requirements.txt
redis
python-dotenv
google-generativeai
pinecone-client
pymupdf   
requests
aiohttp           
numpy 